分離したクラスのテスト追加
